sales-sync-agent/requirements.txt
flask>=3.0
google-api-python-client>=2.100
google-auth>=2.23
google-auth-oauthlib>=1.1
google-auth-httplib2>=0.1
